extract locators to it's own files
