+++ lib/reek/source/locators/stdin.rb
@@ -1,0 +1,20 @@
+require_relative '../source_locator'
+
+module Reek
+  module Source
+    module Locators
+      #
+      # Finds Ruby sources in STDIN
+      #
+      class Stdin < SourceLocator
+        def self.handle?(source)
+          source.is_a?(IO)
+        end
+
+        def locate
+          [source]
+        end
+      end
+    end
+  end
+end
