+++ reek.gemspec
@@ -26,6 +26,7 @@ Gem::Specification.new do |s|
   s.add_runtime_dependency 'private_attr', '~> 1.1'
   s.add_runtime_dependency 'rainbow',      '~> 2.0'
   s.add_runtime_dependency 'unparser',     '~> 0.2.2'
+  s.add_runtime_dependency 'descendants_tracker', '~> 0.0.4'
 
   s.add_development_dependency 'activesupport', '~> 4.2'
   s.add_development_dependency 'aruba',         '~> 0.8.0'
